Export function used by CLI to merge outputs
